gateway/requirements.txt
# FastAPI & Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Temporal Client
temporalio>=1.7.0

# Voice & AI Services
hume>=0.7.0
zep-cloud>=2.0.0
google-generativeai>=0.8.3

# Database
psycopg[binary]>=3.1.0

# HTTP Client (for Hume OAuth)
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging
structlog>=24.1.0

# Memory (SuperMemory for user personalization)
supermemory>=0.1.0
